LeafProject/leaf-regeneration-classification/requirements.txt
Flask==2.0.1
numpy==1.21.0
pandas==1.3.0
scikit-learn==0.24.2
tensorflow==2.5.0
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.1.48
matplotlib==3.4.2
seaborn==0.11.1
Pillow==8.2.0
pyyaml==5.4.1
jupyter==1.0.0
requests==2.25.1